Fix .py files
helper.py:
Removed "NOT NULL" retraint on crYmd (DateCreated) in DB because it is not always present in an article.

application.py:
Made record dates-checking for missing_elements more lenient.
Previously, an article with missing crYmd would not be inserted
into DB. Now, an article with any one of crYmd,
coYmd (DateCompleted) and reYmd (DateRevised) will be inserted
into DB.